Add CI/CD Tools
Add clang-tidy, address sanitizer, undefined sanitizer, thread sanitizer, initial code coverage, and clang-format to the build pipeline

Note: memory sanitizer and leak sanitizer are not added currently since it is required to build all dependent libraries with this flag.  In order to support this, we would have to build GMP from source which is currently not in scope for this work item.